Move keying to CacheDatastore
