+++ library/src/test/java/com/mux/player/CacheControllerTests.kt
@@ -47,41 +47,4 @@ class CacheControllerTests : AbsRobolectricTest() {
     testTheCase(cacheControlValueSimple)
     testTheCase(cacheControlValueConflicting)
   }
-
-  @Test
-  fun `segmentCacheKey generates different keys for segments`() {
-    val notASegmentUrl = "https://manifest-gcp-us-east4-vop1.cfcdn.mux.com/efg456hjk/rendition.m3u8"
-    val hlsSegmentUrl = "https://chunk-gcp-us-east4-vop1.cfcdn.mux.com/v1/chunk/hls123abc/0.ts"
-    val cmafSegmentUrl = "https://chunk-gcp-us-east4-vop1.cfcdn.mux.com/v1/chunk/cmaf456def/146.m4s"
-
-    val notASegmentKey =
-      CacheController.generateCacheKey(URL(notASegmentUrl))
-    val hlsKey = CacheController.generateCacheKey(URL(hlsSegmentUrl))
-    val cmafKey = CacheController.generateCacheKey(URL(cmafSegmentUrl))
-
-    Assert.assertEquals(
-      "Non-segment URLs key on the entire URL",
-      notASegmentUrl, notASegmentKey
-    )
-    Assert.assertNotEquals(
-      "HLS segment URLs have a special key",
-      hlsKey, hlsSegmentUrl
-    )
-    Assert.assertNotEquals(
-      "CMAF segment URLs have a special key",
-      cmafKey, cmafSegmentUrl
-    )
-  }
-
-  @Test
-  fun `segmentCacheKey generates cache keys for segments correctly`() {
-    val segmentUrl = "https://chunk-gcp-us-east4-vop1.cfcdn.mux.com/v1/chunk/abc12345xyz/0.ts"
-    val expectedKey = "/v1/chunk/abc12345xyz/0.ts"
-
-    val key = CacheController.generateCacheKey(URL(segmentUrl))
-    Assert.assertEquals(
-      "cache key should be constructed properly",
-      expectedKey, key
-    )
-  }
 }
